initial commit of reworking
