+++ package.json
@@ -1,12 +1,12 @@
 {
     "author": "Simon Hampton",
     "name": "elm-webpack-starter",
-    "version": "3.11.0",
+    "version": "4.0.0",
     "description": "Elm 0.19 starter with Webpack 5 hot-reloading",
     "main": "index.js",
     "scripts": {
         "test": "elm-test",
-        "start": "npm run dev",
+        "start": "webpack serve",
         "dev": "webpack serve --hot --port 3000",
         "build": "webpack",
         "prod": "webpack --mode production",
@@ -18,12 +18,12 @@
     },
     "license": "MIT",
     "devDependencies": {
-        "@babel/core": "^7.12.3",
-        "@babel/preset-env": "^7.12.1",
-        "babel-loader": "^8.2.1",
+        "@babel/core": "^7.12.10",
+        "@babel/preset-env": "^7.12.11",
+        "babel-loader": "^8.2.2",
         "clean-webpack-plugin": "^3.0.0",
         "closure-webpack-plugin": "^2.3.0",
-        "copy-webpack-plugin": "^6.3.1",
+        "copy-webpack-plugin": "^6.4.1",
         "css-loader": "^5.0.1",
         "elm": "^0.19.1-3",
         "elm-analyse": "^0.16.5",
@@ -33,18 +33,18 @@
         "elm-webpack-loader": "^7.0.1",
         "file-loader": "^6.2.0",
         "google-closure-compiler": "^20201102.0.1",
-        "html-webpack-plugin": "^5.0.0-alpha.14",
-        "mini-css-extract-plugin": "^1.3.1",
+        "html-webpack-plugin": "^5.0.0-beta.1",
+        "mini-css-extract-plugin": "^1.3.3",
         "node-sass": "^5.0.0",
         "optimize-css-assets-webpack-plugin": "^5.0.4",
         "resolve-url-loader": "^3.1.2",
         "sass-loader": "^10.1.0",
         "style-loader": "^2.0.0",
         "url-loader": "^4.1.1",
-        "webpack": "^5.5.0",
+        "webpack": "^5.11.0",
         "webpack-cli": "^4.2.0",
         "webpack-dev-server": "^3.11.0",
-        "webpack-merge": "^5.4.0"
+        "webpack-merge": "^5.7.3"
     },
     "dependencies": {
         "purecss": "^2.0.3"
