+++ src/styles.scss
@@ -1,23 +1,27 @@
-@import '~purecss/build/pure.css';
-
-body {
-    background-color: #ddd;
-    margin-top: 20px;
-}
-.container {
-    max-width: 800px;
-    margin-left: auto;
-    margin-right: auto;
-}
-header {
-    display: flex;
-    align-items: center;
-    img {
-        width: 60px;
-        margin-right: 30px;
-    }
+.icon {
+    background-image: url("./assets/images/logo.png");
+    width: 80px;
+    height: 80px;
+    background-size: cover;
 }
 
-button {
-    margin-right: 15px;
-}
+//body {
+//    background-color: #ddd;
+//    margin-top: 20px;
+//}
+//.container {
+//    max-width: 800px;
+//    margin-left: auto;
+//    margin-right: auto;
+//}
+//header {
+//    display: flex;
+//    align-items: center;
+//    img {
+//        width: 60px;
+//        margin-right: 30px;
+//    }
+//}
+//button {
+//    margin-right: 15px;
+//}
