+++ webpack.config.js
@@ -1,202 +1,54 @@
-const path = require("path");
-const {merge} = require('webpack-merge');
+const path = require('path');
 
-const ClosurePlugin = require("closure-webpack-plugin");
-const CopyWebpackPlugin = require("copy-webpack-plugin");
-const HTMLWebpackPlugin = require("html-webpack-plugin");
-const {CleanWebpackPlugin} = require("clean-webpack-plugin");
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const {CleanWebpackPlugin} = require('clean-webpack-plugin');
 
-// Production CSS assets - separate, minimised file
-const MiniCssExtractPlugin = require("mini-css-extract-plugin");
-const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
+const withDebug = true;
 
-var MODE =
-    process.env.npm_lifecycle_event === "prod" ? "production" : "development";
-var withDebug = !process.env["npm_config_nodebug"] && MODE === "development";
-
-console.log(
-    "\x1b[36m%s\x1b[0m",
-    `** elm-webpack-starter: mode "${MODE}", withDebug: ${withDebug}\n`
-);
-
-var common = {
-    mode: MODE,
-    context: path.resolve(__dirname),
-    entry: "./src/index.js",
-    output: {
-        path: path.join(__dirname, "dist"),
-        publicPath: "/",
-        filename: MODE === "production" ? "[name]-[hash].js" : "index.js"
-    },
+module.exports = {
+    mode: 'development',
+    entry: './src/index.js',
     plugins: [
-        new HTMLWebpackPlugin({
-            // Use this template to get basic responsive meta tags
-            template: "src/index.html",
-            // inject details of output file at end of body
-            inject: "body"
-        })
+        new HtmlWebpackPlugin({
+            title: 'Elm Webpack Starter - Development',
+        }),
+        new CleanWebpackPlugin(),
     ],
+    output: {
+        path: path.resolve(__dirname, 'dist'),
+        filename: 'bundle.js'
+    },
     resolve: {
-        modules: [path.join(__dirname, "src"), "node_modules"],
-        extensions: [".js", ".elm", ".scss", ".png"]
+        modules: [path.join(__dirname, "src"), 'node_modules'],
+        extensions: [".elm", ".js"]
+    },
+    devServer: {
+        contentBase: './dist',
+        port: 3000
     },
     module: {
         rules: [
             {
-                test: /\.js$/,
-                exclude: /node_modules/,
-                use: {
-                    loader: "babel-loader"
-                }
-            },
-            {
-                test: /\.scss$/,
+                test: /\.scss$/i,
+                use: ['style-loader', 'css-loader', "sass-loader"],
+            }, {
+                test: /\.elm$/,
                 exclude: [/elm-stuff/, /node_modules/],
-                // see https://github.com/webpack-contrib/css-loader#url
-                use: ["style-loader", "css-loader?url=false", "sass-loader"]
-            },
-            {
-                test: /\.css$/,
-                exclude: [/elm-stuff/, /node_modules/],
-                use: ["style-loader", "css-loader?url=false"]
-            },
-            {
-                test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-                exclude: [/elm-stuff/, /node_modules/],
-                use: {
-                    loader: "url-loader",
-                    options: {
-                        limit: 10000,
-                        mimetype: "application/font-woff"
-                    }
-                }
-            },
-            {
-                test: /\.(ttf|eot|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-                exclude: [/elm-stuff/, /node_modules/],
-                use: {loader: "file-loader"}
-            },
-            {
-                test: /\.(jpe?g|png|gif|svg)$/i,
-                exclude: [/elm-stuff/, /node_modules/],
-                use: {loader: "file-loader"}
-            }
-        ]
-    }
-};
-
-if (MODE === "development") {
-    module.exports = merge(common, {
-        optimization: {
-            // Prevents compilation errors causing the hot loader to lose state
-            emitOnErrors: false
-        },
-        module: {
-            rules: [
-                {
-                    test: /\.elm$/,
-                    exclude: [/elm-stuff/, /node_modules/],
-                    use: [
-                        {loader: "elm-hot-webpack-loader"},
-                        {
-                            loader: "elm-webpack-loader",
-                            options: {
-                                // add Elm's debug overlay to output
-                                debug: withDebug
-                            }
-                        }
-                    ]
-                }
-            ]
-        },
-        devServer: {
-            inline: true,
-            stats: "errors-only",
-            contentBase: path.join(__dirname, "src/assets"),
-            publicPath: "/",
-            historyApiFallback: true,
-            // feel free to delete this section if you don't need anything like this
-            before(app) {
-                // on port 3000
-                app.get("/test", function (req, res) {
-                    res.json({result: "OK"});
-                });
-            }
-        }
-    });
-}
-
-if (MODE === "production") {
-    module.exports = merge(common, {
-        optimization: {
-            minimizer: [
-                new ClosurePlugin(
-                    {mode: "STANDARD"},
+                use: [
                     {
-                        // compiler flags here
-                        //
-                        // for debugging help, try these:
-                        //
-                        // formatting: 'PRETTY_PRINT',
-                        // debug: true
-                        // renaming: false
-                    }
-                ),
-                new OptimizeCSSAssetsPlugin({})
-            ]
-        },
-        plugins: [
-            // Delete everything from output-path (/dist) and report to user
-            new CleanWebpackPlugin({
-                root: __dirname,
-                exclude: [],
-                verbose: true,
-                dry: false
-            }),
-            // Copy static assets
-            new CopyWebpackPlugin({
-                patterns: [
-                    {
-                        from: "src/assets"
-                    }
-                ]
-            }),
-            new MiniCssExtractPlugin({
-                // Options similar to the same options in webpackOptions.output
-                // both options are optional
-                filename: "[name]-[hash].css"
-            })
-        ],
-        module: {
-            rules: [
-                {
-                    test: /\.elm$/,
-                    exclude: [/elm-stuff/, /node_modules/],
-                    use: {
                         loader: "elm-webpack-loader",
                         options: {
-                            optimize: true
+                            // add Elm's debug overlay to output
+                            debug: withDebug,
+                            optimize: false
                         }
                     }
-                },
-                {
-                    test: /\.css$/,
-                    exclude: [/elm-stuff/, /node_modules/],
-                    use: [
-                        MiniCssExtractPlugin.loader,
-                        "css-loader?url=false"
-                    ]
-                },
-                {
-                    test: /\.scss$/,
-                    exclude: [/elm-stuff/, /node_modules/],
-                    use: [
-                        MiniCssExtractPlugin.loader,
-                        "css-loader?url=false",
-                        "sass-loader"
-                    ]
-                }
-            ]
-        }
-    });
-}
+                ]
+            }, {
+                test: /\.(png|svg|jpg|jpeg|gif)$/i,
+                type: 'asset/resource',
+            },
+        ],
+    },
+};
+
